new cpp Sol

+++ cpp_Solution_3/array_pointer.pxd
@@ -3,5 +3,5 @@ cdef extern from "array_pointer.cpp":
 
 #needed functions for pysol.pyx file
 cdef extern from "array_pointer.h":
-	float* arr_to_ptr(int m)
-	float** arr_to_ptr_ptr(int m)
+	double* arr_to_ptr(int m)
+	double** arr_to_ptr_ptr(int m)
